edit: my page에서 EditProfile 컴포넌트 분리
